DABOM-328 refactor: 파일구조 옮기기
