+++ apps/service/src/components/recap/utils/BalancedOpalescentBackground.tsx
