+++ apps/service/src/components/recap/utils/RecapStep2Background.tsx
